+++ src/Twig/Nodes/HeaderNode.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CraftCms\Cms\Twig\Nodes;
 
-use CraftCms\Cms\Http\Middleware\SetHeaders;
+use CraftCms\Cms\Support\Facades\ResponseHeaders;
 use Override;
 use Twig\Attribute\YieldReady;
 use Twig\Compiler;
@@ -20,7 +20,7 @@ class HeaderNode extends Node
             ->write('$_headerParts = array_map(\'trim\', explode(\':\', ')
             ->subcompile($this->getNode('header'))
             ->raw(", 2));\n")
-            ->write(SetHeaders::class."::add(\$_headerParts[0], \$_headerParts[1] ?? '');\n")
+            ->write(ResponseHeaders::class."::add(\$_headerParts[0], \$_headerParts[1] ?? '');\n")
             ->write("unset(\$_headerParts);\n");
     }
 }